support python3.9 and 3.10, auto detect gpu in some shell script, add update libraries scripts (#576)
* support other python version for dockerfile.base and build_base.sh

* upda Update the README document to include the selection of Python version for the base Docker and remove unnecessary variables from the Dockerfile of the base Docker.

* revise docker readme for docker.base

* revise readme of docker for building base docker for different python version

* revise docker/common/install_math.sh for auto detecting the arch of gpu

* Due to the potential changes in torch_musa's patch and the recompilation of torch during CI, compiling torch in Docker does not provide any benefits, but instead slows down Docker production.

* revise readme in docker

* update_release_mudnn.sh -> update_dependency_mudnn.sh, because the mudnn updated by this script file is not the release version but the relatively stable daily version mudnn

* Revert "Due to the potential changes in torch_musa's patch and the recompilation of torch during CI, compiling torch in Docker does not provide any benefits, but instead slows down Docker production."

This reverts commit 0fdca6515b6496dc6a6ce3ea879e558c5d318d24.

* add  scripts/update_dependency_mudnn.sh

* install torch_musa in dockerfile.dev and run the ut then uninstall it

* fix bug in dockerfile.dev

* python setup.py install ---> bash build.sh -m for torch_musa

* revise scripts/run_daily_release.sh to auto ut and it, and revise jenkinsfile to not release when ut and it failed for python 3.8 3.9 3.10

* Install different underlying libraries based on the results of mthreads-gmi.

* for the same purpose

* Replace mudnn in update_release_mudnn.sh with the release version published by the mudnn team instead of relatively stable daily version.

* add build_all.sh and comment a func that dev1.5.1 mudnn does not support

* fix bug in build_all.sh

* fix bug of Jenkinsfile

* lint

* lint

* fix bug of update_release_mudnn.sh

* check building result in docker/build_all.sh

* auto detect gpu in docker/common/install_math.sh

* add -y for pip uninstall in dockerfile.dev

* Enter the GitHub account password in the script so that you don't have to enter it repeatedly

* Revert "Enter the GitHub account password in the script so that you don't have to enter it repeatedly"

This reverts commit 4c864ae5d0b8b35d31cdffd96bb3cb9b74fcd790.

* ut

* Revert "ut"

This reverts commit 5829ee6d535625bef8e85f2b8fcbbe49141ec96d.

* Restore scripts/update_release_mudnn.sh to the original version

* revise build_all.sh for the download url of mudnn

* revise readme.md in docker folder

* move some script from ./scripts to ./docker/common/

* optimize the experience of build_all.sh

* change the commit id to git clone in build_all.sh

* bug fixed

* fix bug in docker/build.sh

* fix bug in docker/build_all.sh

* bug fixed

* revise docker/README.md

* bug fixed

* will not exit when torch_musa UT failed

* revise docker/build_all.sh for bug

* save the output of ut and integrxxx test in txt

* fix bug

* revise test_linalg_lstsq.py for potential bug

* fix bug in docker/build.sh and docker/ubuntu/dockerfile.dev

* add torch_musa_tag in build.sh build_all.sh

* readme.md

* revise readme.md

* revise build_all.sh

* docker/build_all.sh bug fixed

* revise readme.md

* add docker/common/update_daily_musa_toolkits.sh and revise readme.md for updating libraries

* update README.md

* update readme.md

* revise daily_release.sh

* update docker/ubuntu/dockerfile.dev for no uninstalling torch_musa

* skip ci

* revise readme.md [skip ci]

* set GPU ARCH to qy1 when mthreads-gmi failed

* install pytest for daily release

* revise readme.md

* change ci to sh-harbor.mthreads.com/mt-ai/musa-pytorch-dev-py38:latest

* cancel ut for daily release considering main branch have passed ci

* delete useless code in Jenkinsfile

* recover big tensor input for reduce

* lint

* polish the readme

* delete build folder in dockerfile.dev

* readme

* copy vision to /home

* library -> libraries

* bug

* python -> Python

* dev -> development

* polish readme

* polish

* polish

* polish

* polish

* polish

* polish

* polish

* polish

* polish

* polish

* polish

+++ docker/build.sh
@@ -22,6 +22,8 @@ MCCL_URL=""
 TORCH_WHL_URL=""
 TORCH_MUSA_WHL_URL=""
 VISION_TAG="v0.16.0"
+NO_PREPARE=0
+TORCH_MUSA_TAG='dev1.5.1'
 
 usage() {
   echo -e "\033[1;32mThis script is used to build docker image for torch_musa. \033[0m"
@@ -37,11 +39,13 @@ usage() {
   echo -e "\033[32m    --mccl_url)                : The download link of MCCL. \033[0m"
   echo -e "\033[32m    --torch_whl_url)           : The download link of torch wheel. \033[0m"
   echo -e "\033[32m    --torch_musa_whl_url)      : The download link of torch_musa wheel. \033[0m"
+  echo -e "\033[32m    --torch_musa_tag)          : The tag of torch_musa to be installed \033[0m"
   echo -e "\033[32m    -h/--help                  : Help information. \033[0m"
+  echo -e "\033[32m    --no_prepare               : Whether to prepare build context, such as torch_musa. \033[0m"
 }
 
 # parse parameters
-parameters=$(getopt -o rn:t:b:f:v:m:h:: --long name:,tag:,base_img:,docker_file:,python_verison:,musa_toolkits_url:,mudnn_url:,release,mccl_url:,torch_whl_url:,torch_musa_whl_url:,help::, -n "$0" -- "$@")
+parameters=$(getopt -o rn:t:b:f:v:m:h:: --long name:,tag:,base_img:,docker_file:,python_verison:,musa_toolkits_url:,mudnn_url:,torch_musa_tag:,release,no_prepare,mccl_url:,torch_whl_url:,torch_musa_whl_url:,help::, -n "$0" -- "$@")
 [ $? -ne 0 ] && { echo -e "\033[34mTry '$0 --help' for more information. \033[0m"; exit 1; }
 
 eval set -- "$parameters"
@@ -56,9 +60,11 @@ while true;do
     -m|--musa_toolkits_url) MUSA_TOOLKITS_URL=$2; shift 2;;
     --mudnn_url) MUDNN_URL=$2; shift 2;;
     -r|--release) RELEASE=1; shift ;;
+    --no_prepare) NO_PREPARE=1; shift ;;
     --mccl_url) MCCL_URL=$2; shift 2;;
     --torch_whl_url) TORCH_WHL_URL=$2; shift 2;;
     --torch_musa_whl_url) TORCH_MUSA_WHL_URL=$2; shift 2;;
+    --torch_musa_tag) TORCH_MUSA_TAG=$2; shift 2;;
     -h|--help) usage; exit ;;
     --) shift ; break ;;
     *) usage; exit 1 ;;
@@ -76,7 +82,7 @@ function prepare_build_context() {
       git clone -b ${VISION_TAG}  https://github.com/pytorch/vision.git --depth=1 $TORCH_VISION_ROOT_DIR/vision
     fi
     sudo cp -r $TORCH_VISION_ROOT_DIR/vision $BUILD_DIR
-    sudo git clone https://github.mthreads.com/mthreads/torch_musa.git $BUILD_DIR/torch_musa
+    sudo git -b ${TORCH_MUSA_TAG} clone https://github.mthreads.com/mthreads/torch_musa.git $BUILD_DIR/torch_musa
   fi
   CUR_ROOT=$(cd "$(dirname "$0")"; pwd)
   sudo cp -r $CUR_ROOT/common $BUILD_DIR/
@@ -104,8 +110,12 @@ function build_docker_image() {
 }
 
 BUILD_DIR=${DOCKER_BUILD_DIR:-$(pwd)/tmp}
-sudo mkdir -p $BUILD_DIR
-prepare_build_context ${BUILD_DIR}
+if [ "$NO_PREPARE" -eq 0 ]; then
+  sudo mkdir -p $BUILD_DIR
+  prepare_build_context ${BUILD_DIR}
+fi
 build_docker_image $BUILD_DIR
-sudo rm -rf $BUILD_DIR
+if [ "$NO_PREPARE" -eq 0 ]; then
+  sudo rm -rf $BUILD_DIR
+fi
